Add separate buttons for sign up and login.
There's confusion when people are trying to sign up, and
the default auth0 screen is for login. This tries to solve
the problem by adding distinct buttons for logging in versus
signing up, and based on which is selected displaying a different
screen in the auth0 widget.

This is achieved by (ab)using the `login_hint` request parameter
in the oauth2 flow. We send `login_hint=login` if the user presses
the `login` button and `login_hint=signUp` if they press `signUp`
button. Then we configure the Auth0 Lock widget to render the desired
screen by setting the [`initialScreen` config parameter](https://auth0.com/docs/libraries/lock/v11/configuration#initialscreen-string-)
based on the passed value.

Note - because the actual login widget and configuration is hosted
by Auth0, the changes are not part of this commit. But for those interested,
the change I made was to add the following to the initialization code:
```javascript
initialScreen: config.extraParams.login_hint === 'login' ? 'login' : 'signUp'
```

Fixes #154