Batch approval of pending users
See #182

Implementation:
* A new button in the admin UI
* A route for the new API endpoint
* A service library for making the update.

The service will only approve the list of curently loaded users (i.e.
the bound React data), and only those which are in pending state.

Importantly, the steps an admin must follow:

1. Load the admin page.
2. Manually remove flagged pending accounts (using @sjforman's
incantations).
3. Click on "Batch Approve".

If you refresh the page before step 3, it's possible someone could have
signed up in the interim.

Next steps should be:
1. Add some guard around this button. E.g. a "Confirm I cam Scott Forman
and that I really want to do this" checkbox.
2. Allow the admin to read the submission text and ban accounts (e.g.
what Scott is doing in the DB today).